# apps/api/docs/openapi.yaml
{
  "components": {
    "schemas": {
...
      "PasswordPatchDto": {
        "properties": {},
        "type": "object"
      },
      "ProfilePatchDto": {
        "properties": {},
        "type": "object"
      },
      "SettingsPatchDto": {
        "properties": {},
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "HTTP API for GroceryView products, stores, prices, users, watchlists, baskets, and alerts.",
    "title": "GroceryView API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "AuthenticatedSettingsController_readPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Read authenticated user preferences for currency, preferred stores, notification channels, and MyFlyer algorithm_choice"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "AuthenticatedSettingsController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Save authenticated user preferences for currency, preferred stores, notification channels, and MyFlyer algorithm_choice"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/profile": {
      "get": {
        "operationId": "AuthenticatedSettingsController_readProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Read authenticated user profile metadata including display name, email, and account creation date"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "AuthenticatedSettingsController_updateProfile",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update authenticated user profile display name"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/profile/password": {
      "patch": {
        "operationId": "AuthenticatedSettingsController_updatePassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordPatchDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Change authenticated user password without returning or logging secrets"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "settings"
        ]
      }
    },
...
  },
  "servers": [],
...
}